# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthCredentials {
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface AuthUser {
  id: number;
  email: string;
  createdAt: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface ConnectDbRequest {
  name?: string;
  connectionString: string;
  readOnly?: boolean;
}

export interface UpdateConnectionRequest {
  name?: string;
  readOnly?: boolean;
}

export interface SavedConnection {
  id: number;
  name: string;
  host: string;
  databaseName: string;
  readOnly: boolean;
  createdAt: string;
}

export interface TablesResponse {
  tables: string[];
}

export interface TableStat {
  table: string;
  rowCount: number;
}

export interface TableStatsResponse {
  tables: TableStat[];
}

export interface ColumnInfo {
  name: string;
  dataType: string;
  nullable: boolean;
  hasDefault: boolean;
  isPrimaryKey: boolean;
}

export interface ColumnsResponse {
  table: string;
  columns: ColumnInfo[];
}

export interface RowObject {
  [key: string]: unknown;
}

export interface RowsResponse {
  table: string;
  columns: ColumnInfo[];
  rows: RowObject[];
  page: number;
  pageSize: number;
  totalCount: number;
}

export interface CreateRowRequest {
  values: RowObject;
}

export interface UpdateRowRequest {
  primaryKey: RowObject;
  values: RowObject;
}

export interface DeleteRowRequest {
  primaryKey: RowObject;
}

export interface RowMutationResponse {
  row: RowObject;
}

export interface SavedQuery {
  id: number;
  connectionId: number;
  name: string;
  sql: string;
  createdAt: string;
}

export interface CreateSavedQueryRequest {
  connectionId: number;
  name: string;
  sql: string;
}

export interface QueryResultResponse {
  columns: string[];
  rows: RowObject[];
}

export interface QueryHistoryItem {
  id: number;
  connectionId: number;
  action: string;
  sql: string;
  status: string;
  error?: string | null;
  createdAt: string;
}

export interface RowChangeHistoryItem {
  id: number;
  connectionId: number;
  tableName: string;
  action: string;
  primaryKey?: RowObject;
  before?: RowObject;
  after?: RowObject;
  createdAt: string;
}

export type ConnectionIdParameter = number;

export type TableNameParameter = string;

export type PageParameter = number;

export type PageSizeParameter = number;

/**
 * Search text across row values
 */
export type SearchParameter = string;

/**
 * Column name for a simple value filter
 */
export type FilterColumnParameter = string;

/**
 * Text matched against the selected filter column
 */
export type FilterValueParameter = string;

/**
 * Column name to sort by
 */
export type SortColumnParameter = string;

export type SortDirParameter =
  (typeof SortDirParameter)[keyof typeof SortDirParameter];

export const SortDirParameter = {
  asc: "asc",
  desc: "desc",
} as const;

export type ListTablesParams = {
  connectionId: ConnectionIdParameter;
};

export type ListTableStatsParams = {
  connectionId: ConnectionIdParameter;
};

export type ListColumnsParams = {
  connectionId: ConnectionIdParameter;
  table: TableNameParameter;
};

export type ListRowsParams = {
  connectionId: ConnectionIdParameter;
  table: TableNameParameter;
  /**
   * @minimum 1
   */
  page?: PageParameter;
  /**
   * @minimum 1
   * @maximum 100
   */
  pageSize?: PageSizeParameter;
  /**
   * Search text across row values
   */
  search?: SearchParameter;
  /**
   * Column name for a simple value filter
   */
  filterColumn?: FilterColumnParameter;
  /**
   * Text matched against the selected filter column
   */
  filterValue?: FilterValueParameter;
  /**
   * Column name to sort by
   */
  sortColumn?: SortColumnParameter;
  sortDir?: SortDirParameter;
};

export type ExportTableCsvParams = {
  /**
   * Search text across row values
   */
  search?: SearchParameter;
  /**
   * Column name for a simple value filter
   */
  filterColumn?: FilterColumnParameter;
  /**
   * Text matched against the selected filter column
   */
  filterValue?: FilterValueParameter;
  /**
   * Column name to sort by
   */
  sortColumn?: SortColumnParameter;
  sortDir?: SortDirParameter;
};

export type CreateRowParams = {
  connectionId: ConnectionIdParameter;
  table: TableNameParameter;
};

export type UpdateRowParams = {
  connectionId: ConnectionIdParameter;
  table: TableNameParameter;
};

export type DeleteRowParams = {
  connectionId: ConnectionIdParameter;
  table: TableNameParameter;
};

export type ListSavedQueriesParams = {
  connectionId: ConnectionIdParameter;
};

export type ListQueryHistoryParams = {
  connectionId: ConnectionIdParameter;
};

export type ListRowChangeHistoryParams = {
  connectionId: ConnectionIdParameter;
};